src/requirements.txt
pyodbc>=4.0.30
google-cloud-bigquery>=2.34.4
python-dotenv>=0.19.2
pandas>=1.3.5
tqdm>=4.64.0
pyarrow>=7.0.0
fastparquet>=0.7.1
